requirements.txt
# MINDI_AI 프로젝트 의존성 패키지
# FastAPI 및 웹 서버 관련
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# AI/ML 모델 관련
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
transformers==4.35.0
openai==1.3.0

# 음성 처리 관련
librosa==0.10.1
soundfile==0.12.1
parselmouth==0.4.3
praat-parselmouth==0.4.5

# 데이터 처리 및 분석
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4
scikit-learn==1.3.0
lightgbm==4.1.0

# 이미지 처리
opencv-python==4.8.1.78
Pillow==10.0.1

# HTTP 클라이언트
httpx==0.25.2
requests==2.31.0

# 환경변수 관리
python-dotenv==1.0.0

# 유틸리티
python-dateutil==2.8.2
pathlib2==2.3.7

# 개발 도구 (선택사항)
# jupyter==1.0.0
# matplotlib==3.7.2
# seaborn==0.12.2

# 시스템 관련
psutil==5.9.6
